Defer AudioParam automation when calling rampTo and keep internal state for the latest ramp
Deferring the time ramp seem to provide a more consistent behaviour across browsers and with large lookAhead values. It also removes the necessity of calling the unclearly speced cancelAndHoldAtTime by providing a custom implementation. This is possible because we're in a subset situation of the WebAudio spec where automation event should always be scheduled in a chronological order.