Reintroduce the global 'fwd' object
This commit is the beginning of a huge rollback to the situation where programs are just executable JS files with an implicit global 'fwd' reference as a unique access point to the whole API.
Having to support ES modules has too much drawbacks :
- The server has to bundle the programs (rollup). This phase generates error messages that are hard to work with. Plus, this method won't allow node modules resolution
- The user has to know the correct path to files for each API module/class/function he wants to use
- There can be inconsistencies and corner cases because of the way the code is bundled by webpack for the runner part and by rollup on the program part.
These are just a few.

Now, the code still needs to be cleaned up and the final phase before closing this branch will be to remove rollup.

I've hidden some changes that I'm too lazy to cherry-pick for another commit in this mess. The reverb node now exposes the wet and dry gain nodes. Also, there's a new 'simplex' number utility method to generate simplex noise. It's just a call to the external simple-noise.js package.